LayoutTests/inspector/worker/runtime-basic-subworker-expected.txt

Test for RuntimeAgent in a Subworker.


== Running test suite: Subworker.Runtime.basic
-- Running test case: Main.Runtime.evaluate
Target - Page - passphrase - page-passphrase

-- Running test case: Subworker.Runtime.evaluate
Target - worker-1.js - passphrase - worker-passphrase

-- Running test case: Main.Runtime.RemoteObjectAndPropertyDescriptor.
Target - Page - location and href - Location: inspector/worker/runtime-basic-subworker.html

-- Running test case: Subworker.Runtime.RemoteObjectAndPropertyDescriptor.
Target - worker-1.js - location and href - WorkerLocation: inspector/worker/resources/worker-1.js
